Move run context menu items to separate file
